Put color pallete into seperate theme file
