+++ Git Streak Tracker Widget/Git_Streak_Tracker_Widget.swift
@@ -81,35 +81,6 @@ struct StreakEntry: TimelineEntry {
     let configuration: ConfigurationIntent
 }
 
-struct ColorPallete {
-    static let darkGreen = Color(
-        red: 10.0 / 255.0,
-        green: 22.0 / 255.0,
-        blue: 0.0 / 255.0
-    )
-    static let midGreen = Color(
-        red: 20.0 / 255.0,
-        green: 40.0 / 255.0,
-        blue: 0.0 / 255.0
-    )
-
-    static let darkGray = Color(
-        red: 15.0 / 255.0,
-        green: 15.0 / 255.0,
-        blue: 15.0 / 255.0
-    )
-    static let midGray = Color(
-        red: 25.0 / 255.0,
-        green: 25.0 / 255.0,
-        blue: 25.0 / 255.0
-    )
-    static let midWhite = Color(
-        red: 218.0 / 255.0,
-        green: 218.0 / 255.0,
-        blue: 218.0 / 255.0
-    )
-}
-
 struct Background : View {
     var entry: Provider.Entry
