+++ ThemeConfig.swift
@@ -1,0 +1,43 @@
+//
+//  ThemeConfig.swift
+//  Git Streak Tracker
+//
+//  Created by Samuel Wood on 1/12/23.
+//
+
+import Foundation
+import SwiftUI
+
+struct ColorPallete {
+    static let highlightGreen = Color(
+        red: 5.0 / 255.0,
+        green: 255.0 / 255.0,
+        blue: 0.0 / 255.0
+    )
+    static let darkGreen = Color(
+        red: 10.0 / 255.0,
+        green: 22.0 / 255.0,
+        blue: 0.0 / 255.0
+    )
+    static let midGreen = Color(
+        red: 20.0 / 255.0,
+        green: 40.0 / 255.0,
+        blue: 0.0 / 255.0
+    )
+
+    static let darkGray = Color(
+        red: 15.0 / 255.0,
+        green: 15.0 / 255.0,
+        blue: 15.0 / 255.0
+    )
+    static let midGray = Color(
+        red: 25.0 / 255.0,
+        green: 25.0 / 255.0,
+        blue: 25.0 / 255.0
+    )
+    static let midWhite = Color(
+        red: 218.0 / 255.0,
+        green: 218.0 / 255.0,
+        blue: 218.0 / 255.0
+    )
+}
